@@ -6,35 +6,35 @@
     "start": "ng serve",
     "build": "ng build",
     "buildprod": "ng build --prod",
-    "testwatch": "ng test",
-    "test": "ng test --watch=false",
-    "lint": "ng lint",
+    "testwatch": "npm run lint && ng test",
+    "test": "npm run lint && ng test --watch=false",
     "e2e": "ng e2e",
+    "lint": "ng lint",
     "webdriver-manager": "webdriver-manager"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.9",
-    "@angular/common": "~8.2.9",
-    "@angular/compiler": "~8.2.9",
-    "@angular/core": "~8.2.9",
-    "@angular/forms": "~8.2.9",
-    "@angular/platform-browser": "~8.2.9",
-    "@angular/platform-browser-dynamic": "~8.2.9",
-    "@angular/router": "~8.2.9",
+    "@angular/animations": "~9.0.4",
+    "@angular/common": "~9.0.4",
+    "@angular/compiler": "~9.0.4",
+    "@angular/core": "~9.0.4",
+    "@angular/forms": "~9.0.4",
+    "@angular/platform-browser": "~9.0.4",
+    "@angular/platform-browser-dynamic": "~9.0.4",
+    "@angular/router": "~9.0.4",
     "rxjs": "~6.5.4",
-    "tslib": "^1.11.1",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.25",
-    "@angular/cli": "~8.3.25",
-    "@angular/compiler-cli": "~8.2.9",
-    "@angular/language-service": "~9.0.6",
-    "@types/node": "~13.9.1",
+    "@angular-devkit/build-angular": "~0.900.4",
+    "@angular/cli": "~9.0.4",
+    "@angular/compiler-cli": "~9.0.4",
+    "@angular/language-service": "~9.0.4",
+    "@types/node": "^13.9.1",
     "@types/jasmine": "~3.5.9",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.2.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.4.1",
@@ -45,6 +45,6 @@
     "protractor": "~5.4.3",
     "ts-node": "~8.6.2",
     "tslint": "~5.20.1",
-    "typescript": "~3.5.3"
+    "typescript": "~3.7.5"
   }
 }
