@@ -6,7 +6,7 @@ import {AppComponent} from './app.component';
 import {ArticleCreationComponent} from './article/article-creation.component';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {HomeComponent} from './home/home.component';
-import {HttpClientModule} from "@angular/common/http";
+import {HttpClientModule} from '@angular/common/http';
 import {ArticleListComponent} from './article/article-list.component';
 
 @NgModule({
