@@ -1,16 +1,20 @@
-import {ArticleFormPage} from "./article-form.po";
-import {browser, by, element, ElementFinder, promise} from "protractor";
-import {ArticleListPage} from "./article-list.po";
-import {Browser} from "./browser";
+import {ArticleFormPage} from './article-form.po';
+import {browser, by, element, ElementFinder, promise} from 'protractor';
+import {ArticleListPage} from './article-list.po';
+import {Browser} from './browser';
 
 
 export interface Browsable {
-  navigateToFormPage(): promise.Promise<ArticleFormPage>
+  navigateToFormPage(): promise.Promise<ArticleFormPage>;
 
-  navigateToListPage(): promise.Promise<ArticleListPage>
+  navigateToListPage(): promise.Promise<ArticleListPage>;
 }
 
 export class NavFragment implements Browsable {
+
+  private static findLink(url): ElementFinder {
+    return element(by.css(`app-root nav .nav-link[href="${url}"]`));
+  }
 
   navigateToFormPage(): promise.Promise<ArticleFormPage> {
     const url = '/new';
@@ -34,9 +38,5 @@ export class NavFragment implements Browsable {
       .then(_ => {
         return new ArticleListPage(this);
       });
-  }
-
-  private static findLink(url): ElementFinder {
-    return element(by.css(`app-root nav .nav-link[href="${url}"]`));
   }
 }
