@@ -1,8 +1,8 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
 import {ArticleCreationComponent} from './article/article-creation.component';
-import {HomeComponent} from "./home/home.component";
-import {ArticleListComponent} from "./article/article-list.component";
+import {HomeComponent} from './home/home.component';
+import {ArticleListComponent} from './article/article-list.component';
 
 
 const routes: Routes = [
